backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
asgiref==3.10.0
    # via
    #   django
    #   django-cors-headers
certifi==2025.10.5
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
cryptography==46.0.3
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-database-url==3.0.1
    # via backend (pyproject.toml)
dj-rest-auth==6.0.0
    # via backend (pyproject.toml)
django==5.2.7
    # via
    #   backend (pyproject.toml)
    #   dj-database-url
    #   dj-rest-auth
    #   django-allauth
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-allauth==0.61.1
    # via dj-rest-auth
django-cors-headers==4.9.0
    # via backend (pyproject.toml)
django-environ==0.12.0
    # via backend (pyproject.toml)
django-filter==25.2
    # via backend (pyproject.toml)
djangorestframework==3.16.1
    # via
    #   backend (pyproject.toml)
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.5.1
    # via backend (pyproject.toml)
drf-yasg==1.21.11
    # via backend (pyproject.toml)
idna==3.11
    # via requests
inflection==0.5.1
    # via drf-yasg
numpy==2.3.4
    # via shapely
oauthlib==3.3.1
    # via requests-oauthlib
openrouteservice==2.3.3
    # via backend (pyproject.toml)
packaging==25.0
    # via drf-yasg
pillow==12.0.0
    # via backend (pyproject.toml)
psycopg2==2.9.11
    # via backend (pyproject.toml)
pycparser==2.23
    # via cffi
pyjwt==2.10.1
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
python3-openid==3.2.0
    # via django-allauth
pytz==2025.2
    # via drf-yasg
pyyaml==6.0.3
    # via drf-yasg
requests==2.32.5
    # via
    #   django-allauth
    #   openrouteservice
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via django-allauth
shapely==2.1.2
    # via backend (pyproject.toml)
sqlparse==0.5.3
    # via django
tzdata==2025.2
    # via django
uritemplate==4.2.0
    # via drf-yasg
urllib3==2.5.0
    # via requests
